Refactoring of Output
- Create a BasicOutputs module for all direct commands to Zigate
- Create a ReadAttributes Module for all ReadAttribute related command
- Enable debugLogging accordingly to the new modules
